Split search into a ton of pieces
