+++ cpp/search/searchpuct.cpp
@@ -1,0 +1,9 @@
+#include "../search/search.h"
+
+#include "../search/searchnode.h"
+
+//------------------------
+#include "../core/using.h"
+//------------------------
+
+
